python/method_profile/requirements.txt
annotated-types==0.5.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pip==22.2.1
pydantic==2.3.0
pydantic_core==2.6.3
requests==2.31.0
setuptools==63.2.0
typing_extensions==4.7.1
urllib3==2.0.4
whatap-python==1.4.4
